requirements.txt
python-telegram-bot==21.6
anthropic>=0.40.0
python-dotenv>=1.0.0
requests>=2.31.0
